Part of Issue #421 

With usage-file: 
```
  NAME                                          MONTHLY QTY  UNIT        PRICE   HOURLY COST  MONTHLY COST 
  aws_route53_record.my_record                                                                              
  ├─ Standard queries (first 1B)                      1,000  1M queries  0.4000       0.5479      400.0000  
  ├─ Standard queries (over 1B)                         100  1M queries  0.2000       0.0274       20.0000  
  ├─ Latency based routing queries (first 1B)         1,000  1M queries  0.6000       0.8219      600.0000  
  ├─ Latency based routing queries (over 1B)            200  1M queries  0.3000       0.0822       60.0000  
  ├─ Geo DNS queries (first 1B)                       1,000  1M queries  0.7000       0.9589      700.0000  
  └─ Geo DNS queries (over 1B)                          500  1M queries  0.3500       0.2397      175.0000  
  Total                                                                               2.6781    1,955.0000 
```

Without usage-file:
```
  aws_route53_record.my_record                                                                              
  ├─ Standard queries (first 1B)                          -  1M queries  0.4000            -             -  
  ├─ Latency based routing queries (first 1B)             -  1M queries  0.6000            -             -  
  └─ Geo DNS queries (first 1B)                           -  1M queries  0.7000            -             -  
  Total                                                                                    -             -  
```